Started testing for wave draw debugging and physics alignment.
Noticed some physics abnormalities which may have been the physics functioning but with unseen graphics.

Homing in on parameters to alter to control and change graphics. Resolved waves not drawing full screen but the solution seems incorrect may be another bug elsewhere at root of issue. Need to further tune waves they do not properly adjust for screen rotation yet and have not retested on phone.

git-svn-id: https://laptop-1j1vpkuu/svn/agame@66 da46e613-9826-43f3-9a83-6061620c19e0